+++ include/openpose/filestream/fileStream.hpp
@@ -10,23 +10,43 @@ namespace op
 {
     OP_API DataFormat stringToDataFormat(const std::string& dataFormat);
 
+    // Save custom float format
+    // Example to read it in Python, assuming a (18 x 300 x 500) size Array
+    // x = np.fromfile(heatMapFullPath, dtype=np.float32)
+    // assert x[0] == 3 # First parameter saves the number of dimensions (18x300x500 = 3 dimensions)
+    // shape_x = x[1:x[0]]
+    // assert len(shape_x[0]) == 3 # Number of dimensions
+    // assert shape_x[0] == 18 # Size of the first dimension
+    // assert shape_x[1] == 300 # Size of the second dimension
+    // assert shape_x[2] == 500 # Size of the third dimension
+    // arrayData = x[1+int(round(x[0])):]
+    OP_API void saveFloatArray(const Array<float>& array, const std::string& fullFilePath);
+
     // Save/load json, xml, yaml, yml
-    OP_API void saveData(const std::vector<cv::Mat>& cvMats, const std::vector<std::string>& cvMatNames, const std::string& fileNameNoExtension, const DataFormat format);
+    OP_API void saveData(const std::vector<cv::Mat>& cvMats, const std::vector<std::string>& cvMatNames,
+                         const std::string& fileNameNoExtension, const DataFormat format);
 
-    OP_API void saveData(const cv::Mat& cvMat, const std::string cvMatName, const std::string& fileNameNoExtension, const DataFormat format);
+    OP_API void saveData(const cv::Mat& cvMat, const std::string cvMatName, const std::string& fileNameNoExtension,
+                         const DataFormat format);
 
-    OP_API std::vector<cv::Mat> loadData(const std::vector<std::string>& cvMatNames, const std::string& fileNameNoExtension, const DataFormat format);
+    OP_API std::vector<cv::Mat> loadData(const std::vector<std::string>& cvMatNames,
+                                         const std::string& fileNameNoExtension, const DataFormat format);
 
-    OP_API cv::Mat loadData(const std::string& cvMatName, const std::string& fileNameNoExtension, const DataFormat format);
+    OP_API cv::Mat loadData(const std::string& cvMatName, const std::string& fileNameNoExtension,
+                            const DataFormat format);
 
     // Json - Saving as *.json not available in OpenCV verions < 3.0, this function is a quick fix
-    OP_API void saveKeypointsJson(const Array<float>& keypoints, const std::string& keypointName, const std::string& fileName, const bool humanReadable);
+    OP_API void saveKeypointsJson(const Array<float>& keypoints, const std::string& keypointName,
+                                  const std::string& fileName, const bool humanReadable);
 
     // It will save a bunch of Array<float> elements
-    OP_API void saveKeypointsJson(const std::vector<std::pair<Array<float>, std::string>>& keypointVector, const std::string& fileName, const bool humanReadable);
+    OP_API void saveKeypointsJson(const std::vector<std::pair<Array<float>, std::string>>& keypointVector,
+                                  const std::string& fileName, const bool humanReadable);
 
     // Save/load image
-    OP_API void saveImage(const cv::Mat& cvMat, const std::string& fullFilePath, const std::vector<int>& openCvCompressionParams = {CV_IMWRITE_JPEG_QUALITY, 100, CV_IMWRITE_PNG_COMPRESSION, 9});
+    OP_API void saveImage(const cv::Mat& cvMat, const std::string& fullFilePath,
+                          const std::vector<int>& openCvCompressionParams
+                            = {CV_IMWRITE_JPEG_QUALITY, 100, CV_IMWRITE_PNG_COMPRESSION, 9});
 
     OP_API cv::Mat loadImage(const std::string& fullFilePath, const int openCvFlags = CV_LOAD_IMAGE_ANYDEPTH);