+++ src/openpose/filestream/heatMapSaver.cpp
@@ -34,14 +34,28 @@ namespace op
                 for (auto i = 0u; i < fileNames.size(); i++)
                     fileNames[i] = {fileNameNoExtension + (i != 0 ? "_" + std::to_string(i) : "") + "." + mImageFormat};
 
-                // heatMaps -> cvOutputDatas
-                std::vector<cv::Mat> cvOutputDatas(heatMaps.size());
-                for (auto i = 0u; i < cvOutputDatas.size(); i++)
-                    unrollArrayToUCharCvMat(cvOutputDatas[i], heatMaps[i]);
-
-                // Save each heatMap
-                for (auto i = 0u; i < cvOutputDatas.size(); i++)
-                    saveImage(cvOutputDatas[i], fileNames[i]);
+                // Saving on custom floating type "float". Format it:
+                // First, the number of dimensions of the array.
+                // Next elements: the size of each dimension.
+                // Next: all the elements.
+                if (mImageFormat == "float")
+                {
+                    if (heatMaps.size() > 1)
+                        error("Float only implemented for heatMaps.size() == 1.", __LINE__, __FUNCTION__, __FILE__);
+                    for (auto i = 0u; i < heatMaps.size(); i++)
+                        saveFloatArray(heatMaps[i], fileNames[i]);
+                }
+                // Saving on integer type (jpg, png, etc.)
+                else
+                {
+                    // heatMaps -> cvOutputDatas
+                    std::vector<cv::Mat> cvOutputDatas(heatMaps.size());
+                    for (auto i = 0u; i < cvOutputDatas.size(); i++)
+                        unrollArrayToUCharCvMat(cvOutputDatas[i], heatMaps[i]);
+                    // Save each heatMap
+                    for (auto i = 0u; i < cvOutputDatas.size(); i++)
+                        saveImage(cvOutputDatas[i], fileNames[i]);
+                }
             }
         }
         catch (const std::exception& e)